refactor: clean up server to routes/models/utils/validators only
- Merge config/cosmos.py + repositories/ into single utils/db.py
- Delete config/, repositories/, scripts/ directories
- Update main.py and seed.py imports to use src.utils.db
- Update copilot-instructions.md with new server structure
